+++ src/bp/ui-admin/src/Pages/MyAccount/BotpressAccount/LoginModal.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react'
-import { Redirect } from 'react-router-dom'
 import { Alert, Col, Button, Input, FormGroup, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap'
 
-import { login, register, sendResetPassword, isAuthenticated } from '../../../Auth/licensing'
+import * as licensing from '../../../Auth/licensing'
 
 const DEFAULT_STATE = {
   email: '',
@@ -10,17 +9,33 @@ const DEFAULT_STATE = {
   error: null,
   success: null,
   isLoading: false,
-  isRegistering: false
+  isRegistering: false,
+  showResetPasswordLink: false
 }
+
+const WRONG_PASSWORD_CODE = 'auth/wrong-password'
+const MIN_PW_LENGHT = 6
 
 export default class Login extends Component {
   state = { ...DEFAULT_STATE }
+
+  componentDidMount() {
+    this.setSubmitAction()
+  }
+
+  componentDidUpdate() {
+    this.setSubmitAction()
+  }
+
+  setSubmitAction = () => {
+    this.submitAction = this.state.isRegistering ? this.register : this.login
+  }
 
   login = async () => {
     this.setState({ error: null, success: null, isLoading: true })
 
     try {
-      await login({
+      await licensing.login({
         email: this.state.email,
         password: this.state.password
       })
@@ -29,7 +44,7 @@ export default class Login extends Component {
     } catch (error) {
       this.setState({
         error: error.message,
-        showResetPasswordLink: error.code === 'auth/wrong-password',
+        showResetPasswordLink: error.code === WRONG_PASSWORD_CODE,
         isLoading: false
       })
     }
@@ -39,7 +54,7 @@ export default class Login extends Component {
     this.setState({ error: null, isLoading: true })
 
     try {
-      await register({
+      await licensing.register({
         email: this.state.email,
         password: this.state.password
       })
@@ -57,7 +72,7 @@ export default class Login extends Component {
     this.setState({ error: null, success: null, showResetPasswordLink: false })
 
     try {
-      await sendResetPassword({ email: this.state.email })
+      await licensing.sendResetPassword({ email: this.state.email })
       this.setState({
         success: `An email was sent to ${this.state.email} with instructions on how to reset your password.`
       })
@@ -70,12 +85,11 @@ export default class Login extends Component {
     this.setState({ error: null, isRegistering: true, showResetPasswordLink: false })
   }
 
-  handleInputChanged = event => this.setState({ [event.target.name]: event.target.value })
-  handleInputKeyPress = e => {
-    if (e.key === 'Enter') {
-      this.state.isRegistering ? this.register() : this.login()
-    }
+  isFormValid = () => {
+    return this.formEl && this.formEl.checkValidity() && this.state.password.length >= MIN_PW_LENGHT
   }
+
+  handleInputChanged = event => this.setState({ [event.target.name]: event.target.value })
 
   renderResetPassword() {
     return (
@@ -88,71 +102,67 @@ export default class Login extends Component {
     )
   }
 
+  renderToggleRegister() {
+    return (
+      <div className="registerBtn">
+        <Button size="sm" color="link" onClick={this.setRegisterMode}>
+          Create an account
+        </Button>
+      </div>
+    )
+  }
+
   toggle = () => {
     this.setState({ ...DEFAULT_STATE })
     this.props.toggle()
   }
 
   renderForm = () => {
-    if (isAuthenticated()) {
-      return <Redirect to="/licensing/keys" />
-    }
-
     return (
       <Col>
         {this.state.error && <Alert color="danger">{this.state.error}</Alert>}
         {this.state.success && <Alert color="success">{this.state.success}</Alert>}
-        <FormGroup>
-          <label htmlFor="email">Email Address</label>
-          <Input
-            value={this.state.email}
-            type="text"
-            name="email"
-            id="email"
-            onChange={this.handleInputChanged}
-            onKeyPress={this.handleInputKeyPress}
-          />
-        </FormGroup>
-        <FormGroup>
-          <label htmlFor="password">Password</label>
-          <Input
-            value={this.state.password}
-            type="password"
-            name="password"
-            id="password"
-            onChange={this.handleInputChanged}
-            onKeyPress={this.handleInputKeyPress}
-          />
-        </FormGroup>
-        <div>
-          <Button
-            onClick={this.login}
-            disabled={!this.state.email.length || !this.state.password.length || this.state.isLoading}
-            size="m"
-            color="primary"
-          >
-            {this.state.isRegistering ? 'Register' : 'Login'}
-          </Button>
-
-          {!this.state.isRegistering && (
-            <div className="registerBtn">
-              <Button size="sm" color="link" onClick={this.setRegisterMode}>
-                Create an account
-              </Button>
-            </div>
-          )}
-        </div>
+        <form
+          ref={form => {
+            this.formEl = form
+          }}
+          onSubmit={e => {
+            e.preventDefault()
+            this.submitAction()
+          }}
+          noValidate
+        >
+          <FormGroup>
+            <label htmlFor="email">Email Address</label>
+            <Input value={this.state.email} type="email" name="email" id="email" onChange={this.handleInputChanged} />
+          </FormGroup>
+          <FormGroup>
+            <label htmlFor="password">Password</label>
+            <Input
+              value={this.state.password}
+              type="password"
+              name="password"
+              id="password"
+              onChange={this.handleInputChanged}
+            />
+          </FormGroup>
+          <div>
+            <Button type="submit" disabled={!this.isFormValid()} size="m" color="primary">
+              {this.state.isRegistering ? 'Register' : 'Login'}
+            </Button>
+            {!this.state.isRegistering && this.renderToggleRegister()}
+          </div>
+        </form>
       </Col>
     )
   }
 
   render() {
     const page = this.renderForm()
+    const headerLabel = this.state.isRegistering ? 'Create an account' : 'Login to your account'
     return (
       <Modal isOpen={this.props.isOpen} toggle={this.toggle} size="md">
-        <ModalHeader toggle={this.toggle}>
-          {this.state.isRegistering ? 'Create an account' : 'Login to your account'}
-        </ModalHeader>
+        <ModalHeader toggle={this.toggle}>{headerLabel}</ModalHeader>
         <ModalBody>{page}</ModalBody>
         <ModalFooter>{this.state.showResetPasswordLink && this.renderResetPassword()}</ModalFooter>
       </Modal>