+++ src/bp/ui-admin/src/reducers/license.js
@@ -1,23 +1,38 @@
 import api from '../api'
+import { logout } from '../Auth/licensing'
 
 export const FETCH_LICENSE_RECEIVED = 'license/FETCH_LICENSE_RECEIVED'
 export const FETCH_LICENSING_RECEIVED = 'license/FETCH_LICENSING_RECEIVED'
 export const FETCH_KEYS_REQUESTED = 'license/FETCH_KEYS_REQUESTED'
 export const FETCH_KEYS_RECEIVED = 'license/FETCH_KEYS_RECEIVED'
 export const UPDATE_LICENSING_ACCOUNT = 'license/UPDATE_LICENSING_ACCOUNT'
+export const FETCH_PRODUCTS_REQUESTED = 'license/FETCH_PRODUCTS_REQUESTED'
 export const FETCH_PRODUCTS_RECEIVED = 'license/FETCH_PRODUCTS_RECEIVED'
+export const LOGOUT_USER_FROM_LICENSE_SERVER = 'license/LOGOUT_USER_FROM_LICENSE_SERVER'
+export const LICENSE_KEY_UPDATED = 'license/LICENSE_KEY_UPDATED'
 
 const initialState = {
   license: null,
   licensing: null,
-  keys: [],
-  products: [],
-  isLoadingKeys: false,
+  keys: null,
+  products: null,
+  fetchingKeys: false,
+  fetchingProducts: false,
   licensingAccount: null
 }
 
 export default (state = initialState, action) => {
   switch (action.type) {
+    case LICENSE_KEY_UPDATED:
+      const idx = state.keys.findIndex(k => k.stripeSubscriptionId === action.license.stripeSubscriptionId)
+      return {
+        ...state,
+        keys: [...state.keys.slice(0, idx), action.license, ...state.keys.slice(idx + 1)]
+      }
+    case LOGOUT_USER_FROM_LICENSE_SERVER:
+      return {
+        ...initialState
+      }
     case FETCH_LICENSE_RECEIVED:
       return {
         ...state,
@@ -33,14 +48,14 @@ export default (state = initialState, action) => {
     case FETCH_KEYS_REQUESTED:
       return {
         ...state,
-        isLoadingKeys: true
+        fetchingKeys: true
       }
 
     case FETCH_KEYS_RECEIVED:
       return {
         ...state,
         keys: action.keys,
-        isLoadingKeys: false
+        fetchingKeys: false
       }
 
     case UPDATE_LICENSING_ACCOUNT:
@@ -49,14 +64,33 @@ export default (state = initialState, action) => {
         licensingAccount: action.account
       }
 
+    case FETCH_PRODUCTS_REQUESTED:
+      return {
+        ...state,
+        fetchingProducts: true
+      }
+
     case FETCH_PRODUCTS_RECEIVED:
       return {
         ...state,
-        products: action.products
+        products: action.products,
+        fetchingProducts: false
       }
 
     default:
       return state
+  }
+}
+
+export const licenseUpdated = license => ({
+  type: LICENSE_KEY_UPDATED,
+  license
+})
+
+export const logoutUser = () => {
+  logout()
+  return {
+    type: LOGOUT_USER_FROM_LICENSE_SERVER
   }
 }
 
@@ -85,8 +119,8 @@ export const fetchLicensing = () => {
 export const fetchAllKeys = () => async dispatch => {
   dispatch({ type: FETCH_KEYS_REQUESTED })
 
-  const { data } = await api.getLicensing().get('/me/keys')
-
+  const licensing = await api.getLicensing()
+  const { data } = await licensing.get('/me/keys')
   dispatch({
     type: FETCH_KEYS_RECEIVED,
     keys: data.licenses
@@ -94,7 +128,9 @@ export const fetchAllKeys = () => async dispatch => {
 }
 
 export const fetchProducts = () => async dispatch => {
-  const { data } = await api.getLicensing().get(`/prices`)
+  dispatch({ type: FETCH_PRODUCTS_REQUESTED })
+  const licensing = await api.getLicensing()
+  const { data } = await licensing.get(`/prices`)
 
   dispatch({
     type: FETCH_PRODUCTS_RECEIVED,